Use OpenFBX for loading in FBX loader (#263)
Summary:
Pull Request resolved: https://github.com/facebookresearch/momentum/pull/263

## Context

Now that (I believe) the FBX loaders in Momentum with OpenFBX and FBXSDK are identical for what we want to achieve (we have tests for this), it would be good to consolidate them into one.

Eventually, in the upper stack of this Diff, we merge io_openfbx into io_fbx, where we use OpenFBX for loading and FBXSDK for saving. This way the users no longer need to know which one to use with different API.

## This Diff

Make io_fbx to use OpenFBX for loading, removing its own implementation

Reviewed By: EscapeZero

Differential Revision: D72575926

fbshipit-source-id: 51f28231b8553b3e157df3bfa9101142703a5cf1