Rename openfbx_io to openfbx_loader (#267)
Summary:
Pull Request resolved: https://github.com/facebookresearch/momentum/pull/267

## Context

Now that (I believe) the FBX loaders in Momentum with OpenFBX and FBXSDK are identical for what we want to achieve (we have tests for this), it would be good to consolidate them into one.

Eventually, in the upper stack of this Diff, we merge io_openfbx into io_fbx, where we use OpenFBX for loading and FBXSDK for saving. This way the users no longer need to know which one to use with different API.

## This Diff

Rename `openfbx_io.h/cpp` to `openfbx_loader.h/cpp`, in the sense that `X_io.h` is reserved for the public header for parsing X format.

Reviewed By: juliencbmeta

Differential Revision: D72586399

fbshipit-source-id: 125bcd97434ad720194092777e06f16a24925600