Merge openfbx into fbx (#266)
Summary:
Pull Request resolved: https://github.com/facebookresearch/momentum/pull/266

## Context

Now that (I believe) the FBX loaders in Momentum with OpenFBX and FBXSDK are identical for what we want to achieve (we have tests for this), it would be good to consolidate them into one.

Eventually, in the upper stack of this Diff, we merge io_openfbx into io_fbx, where we use OpenFBX for loading and FBXSDK for saving. This way the users no longer need to know which one to use with different API.

## This Diff

Merge `io_openfbx` into `io_fbx` (without changing files names of headers and sources)

Reviewed By: juliencbmeta

Differential Revision: D72584789

fbshipit-source-id: d56e139be194c28fe0b62d7eeaa1045496ac33f0