Temporal smoothing option for the sequence solver (#283)
Summary:
Pull Request resolved: https://github.com/facebookresearch/momentum/pull/283

Adding temporal smoothing parameter to PyMomentum SolverOptions for the SequenceSolver to unblock several projects.

This parameters is used as a weight of ModelParametersSequenceErrorFunction added to the SequenceSolver when setting up the optimization problem.

While this solves our current issues, it does not expose all possible functionality. One may want to, for example, set different smoothing weights for different parameters. A proper solution would be to expose ModelParametersSequenceErrorFunction to PyMomentum. This will likely require implementing TensorSequenceErrorFunction and update interface of solveTensorSequenceIKProblem, buildSequenceSolverFunction functions and all corresponding bindings.

Reviewed By: jeongseok-meta

Differential Revision: D72410749

fbshipit-source-id: dcd691cd7eb197792ed172fe6af6babf2108962c